int and uint256 for _bifB

+++ contracts/BifChain.sol
@@ -3,10 +3,11 @@ pragma solidity >=0.4.22;
 
 contract BifChain {
 
-    mapping(address=> uint256) uniqueId;
+    mapping(address=> int) uniqueId;
     mapping(address => Item[]) List; 
     mapping(address=> bool) registered;
-    uint256 amountOfUser = 0;
+    int amountOfUser = 0;
+    
     address public owner;
 
     constructor() public {
@@ -15,27 +16,28 @@ contract BifChain {
     }
 
     struct Item {
-        uint256 _bifB;
+        // _bifB = (B.) -1 , 0 , 1
+        int _bifB;
         address _address;
         string _description;
         string _bifData;
         uint256 _currentTimeStamp;
     }
 
-    function returnAmountOfUser() public view returns (uint256) {
+    function returnAmountOfUser() public view returns (int) {
         return amountOfUser;
     } 
 
-    function record(uint256 _x,address _y,string memory _t ,string memory _z) public {
+    function record(int _x,address _y,string memory _t ,string memory _z) public {
         List[msg.sender].push(Item(_x,_y,_t,_z,block.timestamp));
     }
 
-    function getRecord(uint256 index) public view returns(uint256 ,address,string memory,string memory,uint256) {
+    function getRecord(uint256 index) public view returns(int ,address,string memory,string memory,uint256) {
         // return Struct Item
     return (List[msg.sender][index]._bifB,List[msg.sender][index]._address,List[msg.sender][index]._description,List[msg.sender][index]._bifData,List[msg.sender][index]._currentTimeStamp); 
     }
 
-     function getRecordBifNoX(uint256 index) public view returns (uint256 _x){
+     function getRecordBifNoX(uint256 index) public view returns (int _x){
         _x = List[msg.sender][index]._bifB;
         return _x;
     }
@@ -54,7 +56,6 @@ contract BifChain {
         _t = List[msg.sender][index]._currentTimeStamp;
         return _t;
     }
-
 
     function getRecordDataD(uint256 index) public view returns (string memory _d){
         _d = List[msg.sender][index]._bifData;
